feat(commands.js): add functionality to dynamically load commands from subdirectories
The `commands.js` file has been added to introduce a new feature that allows dynamically loading commands from subdirectories. This is achieved by using the `fs` and `path` modules from Node.js.

The `getAllSubdirectories` function is used to recursively retrieve all subdirectories within the `commands` directory. Each subdirectory is then checked for JavaScript files with the `.js` extension, and if found, the command is required and added to the `commands` array.

This new functionality enhances the extensibility of the application by allowing the addition of new commands without modifying the main codebase.